Consolidate community menu items
