+++ app/facets/wagtail_hooks.py
@@ -1,0 +1,5 @@
+from wagtail.contrib.modeladmin.options import (
+    ModelAdmin,
+    ModelAdminGroup,
+    modeladmin_register,
+)
